메모개수 상한 수정
