+++ lib/views/list/dialog/nendo_add_memo_dialog.dart
@@ -26,7 +26,7 @@ class _NendoAddMemoDialogState extends State<NendoAddMemoDialog> {
         mainAxisSize: MainAxisSize.min,
         children: [
           const Text(
-            "메모는 8글자 내외로 최대 4개 등록 가능합니다.",
+            "메모는 20글자 내외로 최대 8개 등록 가능합니다.",
             style: TextStyle(
               fontSize: 12.0,
             ),
@@ -103,16 +103,16 @@ class _NendoAddMemoDialogState extends State<NendoAddMemoDialog> {
     int memoCount = nendoController.getNendoData(widget.num).memo?.length ?? 0;
     // 전체공백 제거 + 콤마로 파싱 + 파싱된것중 빈문자열 제거
     List<String> memoList = text.replaceAll('\n', '').replaceAll(' ', '').split(",").where((element) => element != "").toList();
-    if (memoList.length + memoCount > 4) {
+    if (memoList.length + memoCount > 8) {
       setState(() {
-        errorMessage = "메모 개수가 4개가 넘습니다.";
+        errorMessage = "메모 개수가 8개를 넘습니다.";
       });
       return;
     }
     for (String memo in memoList) {
-      if (memo.length > 8) {
+      if (memo.length > 20) {
         setState(() {
-          errorMessage = "8글자를 넘는 메모가 있습니다.";
+          errorMessage = "20글자를 넘는 메모가 있습니다.";
         });
         return;
       }
